requirements.txt
python==3.13.1
PyPDF2==2.11.2
PyMuPDF==1.21.1
pytesseract==0.3.9
Pillow==9.0.1
pdf2image==1.16.3